Fixed React Setup and App.jsx routes
