+++ front-end/src/pages/home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Home() {
+  return (
+    <div>
+      <h1>Home Page</h1>
+      <p>Welcome to the Home Page!</p>
+    </div>
+  );
+}
+
+export default Home;
